backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
bleach>=6.0.0,<7.0.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
channels>=4.0.0,<5.0.0
channels-redis>=4.2.0,<5.0.0
charset-normalizer==3.4.1
coverage==7.7.1
python-dateutil>=2.8.0,<3.0.0
cryptography==44.0.2
daphne
distlib==0.3.9
dj-database-url>=2.1.0,<3.0.0
Django>=5.1.7,<5.2.0
django-cors-headers>=4.3.1,<5.0.0
django-cryptography>=1.1,<2.0.0
django-inmemorystorage==0.1.1
django-rest-knox>=4.2.0,<5.0.0
django-sslserver>=0.22,<1.0
django-stubs==5.2.0
djangorestframework>=3.14.0,<4.0.0
djangorestframework-simplejwt>=5.3.1,<6.0.0
djangorestframework-stubs==3.16.0
drf-yasg>=1.21.7,<2.0.0
filelock==3.18.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
h11==0.14.0
httplib2==0.22.0
identify==2.6.9
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
mypy>=1.5.1,<2.0.0
nodeenv==1.9.1
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.2
pillow>=10.1.0,<11.0.0
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.2.0
proto-plus==1.26.1
protobuf==6.30.2
psycopg2-binary>=2.9.9,<3.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyotp>=2.9.0,<3.0.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv>=1.0.0,<2.0.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
ruff>=0.1.5,<1.0.0
selenium==4.30.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
stripe>=7.0.0,<8.0.0
trio==0.29.0
trio-websocket==0.12.0
types-requests==2.31
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.30.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
